+++ yp-online/src/main/java/kr/co/yourplanet/online/business/user/controller/MemberController.java
@@ -4,9 +4,9 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import kr.co.yourplanet.core.enums.StatusCode;
-import kr.co.yourplanet.online.business.user.dto.MemberDetail;
-import kr.co.yourplanet.online.business.user.dto.MemberFullInfo;
-import kr.co.yourplanet.online.business.user.dto.update.MemberUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.response.MemberDetail;
+import kr.co.yourplanet.online.business.user.dto.response.MemberFullInfo;
+import kr.co.yourplanet.online.business.user.dto.request.MemberUpdateForm;
 import kr.co.yourplanet.online.business.user.service.MemberQueryService;
 import kr.co.yourplanet.online.business.user.service.MemberUpdateService;
 import kr.co.yourplanet.online.common.ResponseForm;
