+++ yp-online/src/main/java/kr/co/yourplanet/online/business/user/dto/request/CreatorJoinForm.java
@@ -1,20 +1,18 @@
-package kr.co.yourplanet.online.business.user.dto.update;
+package kr.co.yourplanet.online.business.user.dto.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import kr.co.yourplanet.core.enums.GenderType;
 import kr.co.yourplanet.core.validation.annotation.ValidEnum;
-import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-@Builder
-public class CreatorUpdateForm {
+public class CreatorJoinForm {
 
     @ValidEnum(enumClass = GenderType.class)
     private GenderType genderType;
 
     @Valid
-    @NotNull(message = "정산 정보는 null일 수 없습니다.")
-    private SettlementForm settlementForm;
+    @NotNull(message = "인스타그램 계정 정보는 null일 수 없습니다.")
+    private InstagramForm instagramForm;
 }
