+++ yp-online/src/main/java/kr/co/yourplanet/online/business/user/dto/request/MemberValidateForm.java
@@ -1,4 +1,4 @@
-package kr.co.yourplanet.online.business.user.dto;
+package kr.co.yourplanet.online.business.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
