+++ yp-online/src/main/java/kr/co/yourplanet/online/business/user/dto/request/SettlementForm.java
@@ -1,4 +1,4 @@
-package kr.co.yourplanet.online.business.user.dto.update;
+package kr.co.yourplanet.online.business.user.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
