+++ yp-online/src/main/java/kr/co/yourplanet/online/business/user/dto/response/MemberFullInfo.java
@@ -1,4 +1,4 @@
-package kr.co.yourplanet.online.business.user.dto;
+package kr.co.yourplanet.online.business.user.dto.response;
 
 import java.time.LocalDate;
