+++ yp-online/src/main/java/kr/co/yourplanet/online/business/user/service/MemberUpdateService.java
@@ -12,10 +12,10 @@ import kr.co.yourplanet.core.enums.BusinessType;
 import kr.co.yourplanet.core.enums.FileType;
 import kr.co.yourplanet.core.enums.MemberType;
 import kr.co.yourplanet.online.business.file.service.FileService;
-import kr.co.yourplanet.online.business.user.dto.update.BaseUpdateForm;
-import kr.co.yourplanet.online.business.user.dto.update.CreatorUpdateForm;
-import kr.co.yourplanet.online.business.user.dto.update.MemberUpdateForm;
-import kr.co.yourplanet.online.business.user.dto.update.SettlementForm;
+import kr.co.yourplanet.online.business.user.dto.request.BaseUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.request.CreatorUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.request.MemberUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.request.SettlementForm;
 import kr.co.yourplanet.online.business.user.repository.MemberRepository;
 import kr.co.yourplanet.online.common.exception.BadRequestException;
 import lombok.RequiredArgsConstructor;
