+++ yp-online/src/test/java/kr/co/yourplanet/stub/MemberFormBuilder.java
@@ -4,11 +4,11 @@ import java.time.LocalDate;
 
 import kr.co.yourplanet.core.enums.BusinessType;
 import kr.co.yourplanet.core.enums.GenderType;
-import kr.co.yourplanet.online.business.user.dto.BusinessForm;
-import kr.co.yourplanet.online.business.user.dto.update.BaseUpdateForm;
-import kr.co.yourplanet.online.business.user.dto.update.CreatorUpdateForm;
-import kr.co.yourplanet.online.business.user.dto.update.MemberUpdateForm;
-import kr.co.yourplanet.online.business.user.dto.update.SettlementForm;
+import kr.co.yourplanet.online.business.user.dto.request.BusinessForm;
+import kr.co.yourplanet.online.business.user.dto.request.BaseUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.request.CreatorUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.request.MemberUpdateForm;
+import kr.co.yourplanet.online.business.user.dto.request.SettlementForm;
 
 public class MemberFormBuilder {
